Fix logplus not working with Cinch 2.2.5.
The monkeypatch used for catching outgoing messages failed to work.
Instead now use the official API of Cinch’s logging framework, which
should be more stable.

Using Cinch’s logging framework for everything is not possible,
because it lacks the information this logger requires: Who sent a
message? Which channel was it sent to? Etc. Cinch’s logging framework
only copes with bare string messages, which is not sufficient unless
one wants re-parse everything via the :incoming and :outgoing methods
(bare IRC commands as sent over the wire), which does not seem right
either.