Remove explicit requires.
Reason for these was that I ran my IRC but in a chroot, which broke
autoload. This is a very specific usecase and should not be in this
general repository. Who runs his bot in a chroot should do these
requires himself and not have this library do it. Newer versions of
thin also changed what exactly needs to be required, breaking the
http_server plugin.