Fix logplus plugin logging PMs by the bot.
This can be a privacy problem. If the bot replies with confident data,
the data would show up in the potentically public HTML log file. This
commit restricts logplus' logging of the bot's messages to what the
bot actually says in public.

Messages of users *to* the bot were never logged (because they don't
generate the :outgoing event logplus hooks into).